Refactor SemanticTextInferenceFieldsIT into an abstract base class
